# pkg/openapi/spec.yaml
openapi: "3.0.3"
info:
  title: TVProxy API
  description: IPTV stream management and proxy server. Manages channels, streams (M3U and SAT>IP), EPG data, and emulates HDHomeRun devices for Plex/Emby/Jellyfin.
  version: "1.0.0"
servers:
  - url: http://localhost:8080
...
paths:
...
  /api/auth/login:
    post:
      summary: Login
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  refresh_token:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/refresh:
    post:
      summary: Refresh access token
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token:
                  type: string
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/invite/{token}:
    parameters:
      - name: token
        in: path
        required: true
        schema:
          type: string
    post:
      summary: Accept an invitation and set password
      tags: [Auth]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Account activated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/auth/logout:
    post:
      summary: Logout
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/auth/me:
    get:
      summary: Get current user
      tags: [Auth]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current user info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ── Users ───────────────────────────────────────────────────────────
  /api/users:
    get:
      summary: List users
      tags: [Users]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      summary: Create user
      tags: [Users]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
                is_admin:
                  type: boolean
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/users/invite:
    post:
      summary: Create invite for a new user
      tags: [Users]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username]
              properties:
                username:
                  type: string
      responses:
        "201":
          description: Invite created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/users/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Get user
      tags: [Users]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update user
      tags: [Users]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                is_admin:
                  type: boolean
      responses:
        "200":
          description: User updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete user
      tags: [Users]
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: User deleted

  # ── M3U Accounts ────────────────────────────────────────────────────
  /api/m3u/accounts:
    get:
      summary: List M3U accounts
      tags: [M3U]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/M3UAccount"
    post:
      summary: Create M3U account (admin only)
      tags: [M3U]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/M3UAccountInput"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/M3UAccount"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/m3u/accounts/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Get M3U account
      tags: [M3U]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Account details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/M3UAccount"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update M3U account (admin only)
      tags: [M3U]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/M3UAccountInput"
      responses:
        "200":
          description: Account updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/M3UAccount"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete M3U account (admin only)
      tags: [M3U]
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Account deleted
  /api/m3u/accounts/{id}/refresh:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      summary: Refresh M3U account streams (async, admin only)
      tags: [M3U]
      security: [{ bearerAuth: [] }]
      responses:
        "202":
          description: Refresh started
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/m3u/accounts/{id}/status:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Get M3U account refresh status
      tags: [M3U]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Refresh status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RefreshStatus"

  # ── Streams ─────────────────────────────────────────────────────────
  /api/streams:
    get:
      summary: List streams
      description: Returns summaries by default. Use ?full=true for all fields. Use ?account_id={id} to filter by M3U account.
      tags: [Streams]
      security: [{ bearerAuth: [] }]
      parameters:
        - name: full
          in: query
          schema:
            type: string
            enum: ["true"]
          description: If "true", returns full stream objects instead of summaries
        - name: account_id
          in: query
          schema:
            type: string
          description: Filter by M3U account ID
      responses:
        "200":
          description: List of streams or stream summaries
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: "#/components/schemas/Stream"
                  - type: array
                    items:
                      $ref: "#/components/schemas/StreamSummary"
  /api/streams/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Get stream
      tags: [Streams]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Stream details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Stream"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete stream (admin only)
      tags: [Streams]
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Stream deleted

  # ── Channels ────────────────────────────────────────────────────────
  /api/channels:
    get:
      summary: List channels
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
    post:
      summary: Create channel
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "201":
          description: Channel created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "400":
          $ref: "#/components/responses/BadRequest"
  /api/channels/{id}:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Get channel
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Channel details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      summary: Update channel
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChannelInput"
      responses:
        "200":
          description: Channel updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      summary: Delete channel
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Channel deleted
  /api/channels/{id}/streams:
    parameters:
      - $ref: "#/components/parameters/id"
    get:
      summary: Get streams assigned to channel
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: List of channel stream assignments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChannelStream"
    post:
      summary: Assign streams to channel
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                stream_ids:
                  type: array
                  items:
                    type: string
      responses:
        "200":
          description: Streams assigned
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /api/channels/{id}/fail:
    parameters:
      - $ref: "#/components/parameters/id"
    post:
      summary: Increment channel fail count
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
    delete:
      summary: Reset channel fail count
      tags: [Channels]
      security: [{ bearerAuth: [] }]
      responses:
        "204":
          description: Fail count reset

  # ── Channel Groups ──────────────────────────────────────────────────
...
  /api/settings/soft-reset:
    post:
      summary: Soft reset (clears user data, keeps system config)
      tags: [Settings]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Soft reset complete
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  parameters:
    id:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Not authenticated
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
        is_admin:
          type: boolean
        invite_token:
          type: string
          nullable: true
        invite_expires_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...